Move replay to RLGym conversion
